os/signal: don't let SIGUSR1 kill the test process in TestStop
CL 18229 added SIGUSR1 to TestStop and updated the signal testing
loop to only send SIGHUP and SIGWINCH outside the os.Notify/os.Stop
pair. However, the update only updated the first such send while
leaving the second. Consequently, an unhandled SIGUSR1 is raised
by TestStop which kills the test binary on darwin/arm{,arm64}.

Uppdate the second send, fixing the darwin/arm{,arm64} test crash
in os/signal.

For #14318

+++ src/os/signal/signal_test.go
@@ -214,7 +214,7 @@ func TestStop(t *testing.T) {
 		// Send the signal.
 		// If it's SIGWINCH, we should not see it.
 		// If it's SIGHUP, maybe we'll die. Let the flag tell us what to do.
-		if sig != syscall.SIGHUP || *sendUncaughtSighup == 2 {
+		if sig == syscall.SIGWINCH || (sig == syscall.SIGHUP && *sendUncaughtSighup == 2) {
 			syscall.Kill(syscall.Getpid(), sig)
 		}
